Develop

+++ server/src/seeds/cleanDb.ts
@@ -3,13 +3,18 @@ import db from '../config/connection.js';
 
 export default async (modelName: "Question", collectionName: string) => {
   try {
-    let modelExists = await models[modelName].db.db.listCollections({
-      name: collectionName
-    }).toArray()
+    const ref = models[modelName].db.db;
 
-    if (modelExists.length) {
-      await db.dropCollection(collectionName);
+    if (ref) {
+      let modelExists = await ref.listCollections({
+        name: collectionName
+      }).toArray()
+  
+      if (modelExists.length) {
+        await db.dropCollection(collectionName);
+      }
     }
+
   } catch (err) {
     throw err;
   }
